% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_infilling.R
\name{do_infilling}
\alias{do_infilling}
\title{Perform Rainfall Data Infilling}
\usage{
do_infilling(
  data,
  station,
  date,
  rainfall,
  rfe,
  metadata = NULL,
  metadata_station = NULL,
  lon,
  lat,
  stations_to_include = NULL,
  rainfall_estimate_column,
  autobins = TRUE,
  by_month = TRUE,
  n_bins = 5,
  custom_bins = c(1, 3, 5, 10, 15, 20),
  count_filter = 10,
  min_rainy_days_threshold = 50,
  target_months = 5:9,
  distribution_flag = c("gamma", "lognormal"),
  markovflag = TRUE,
  b0 = -0.2,
  b1 = 0.05,
  a0 = 0.1,
  a1 = 1,
  b0_dryday = -0.2,
  b0_rainyday = -0.2,
  b1_dryday = 0.05,
  b1_rainyday = 0.05,
  kappa = 2,
  theta = 2,
  p0 = 0.001,
  p0_rainyday = 0.001,
  p0_dryday = 0.001,
  set_seed = NULL,
  return_type = c("numeric", "data.frame")
)
}
\arguments{
\item{data}{Either a path to a CSV file containing historical rainfall data
or a data frame.}

\item{station}{A string specifying the column name in \code{data} that contains
the station names.}

\item{date}{A string specifying the column name in \code{data} that contains
the date values.}

\item{rainfall}{A string specifying the column name in \code{data} that contains
the original rainfall values.}

\item{rfe}{A string specifying the column name in \code{data} that contains
rainfall estimates (RFE) used as predictors for rainfall generation.}

\item{metadata}{An optional data frame containing additional metadata to merge
with the historical data. Default is \code{NULL}.}

\item{metadata_station}{A string specifying the column name in the metadata
data frame that corresponds to \code{station}.
If \code{NULL}, it defaults to the value of \code{station}.}

\item{lon}{A string specifying the column corresponding to longitude values.
If \code{metadata} is \code{NULL}, this column is from \code{data}, otherwise
this is from \code{metadata}}

\item{lat}{A string specifying the column corresponding to latitude values.
If \code{metadata} is \code{NULL}, this column is from \code{data}, otherwise
this is from \code{metadata}}

\item{stations_to_include}{A string specifying the station to include from
the calibration data (Optional). Default is \code{NULL},
meaning that all stations are included.}

\item{rainfall_estimate_column}{A string specifying the column name in \code{data}
that contains rainfall estimates used for
calibration.}

\item{autobins}{Logical. If \code{TRUE}, bins for RFE are computed automatically using quantiles. Default \code{TRUE}.}

\item{by_month}{Logical. If \code{TRUE} and \code{autobins} is also \code{TRUE}, bins are computed separately for each month. Default \code{TRUE}.}

\item{n_bins}{Integer. The number of bins to create when \code{autobins = TRUE}. Default is \code{5}}

\item{custom_bins}{Numeric vector. Used when \code{autobins = FALSE} to specify the RFE bin edges manually.
Default is \code{c(1, 3, 5, 10, 15, 20)}.}

\item{count_filter}{A numeric threshold specifying the minimum number of values
required in each bin to include in calculations. Default is
\code{10}.}

\item{min_rainy_days_threshold}{A numeric threshold for the minimum number of
rainy days required for inclusion in calculations.
Default is \code{50}.}

\item{target_months}{A numeric vector specifying the months (as integers) to
apply dry season parameters. Default is \code{5:9} (May to
September).}

\item{distribution_flag}{A character string specifying the statistical distribution
to use for rainfall generation. Options are \code{"gamma"}
(default) or \code{"lognormal"}.}

\item{markovflag}{A logical value indicating whether to use a Markov chain
approach for rainfall occurrence. Default is \code{TRUE}.}

\item{b0}{A numeric value for the intercept in the dry season model. Default is
\code{-0.2}.}

\item{b1}{A numeric value for the slope in the dry season model. Default is \code{0.05}.}

\item{a0}{A numeric value for the intercept in the rainfall distribution model.
Default is \code{0.1}.}

\item{a1}{A numeric value for the slope in the rainfall distribution model.
Default is \code{1}.}

\item{b0_dryday}{A numeric value for the intercept for dry days. Default is
\code{-0.2}.}

\item{b0_rainyday}{A numeric value for the intercept for rainy days. Default is
\code{-0.2}.}

\item{b1_dryday}{A numeric value for the slope for dry days. Default is \code{0.05}.}

\item{b1_rainyday}{A numeric value for the slope for rainy days. Default is \code{0.05}.}

\item{kappa}{A numeric value specifying the scaling factor for the rainfall
distribution variance. Default is \code{2}.}

\item{theta}{A numeric value specifying the exponent for the rainfall
distribution variance. Default is \code{2}.}

\item{p0}{A numeric value for the baseline probability of precipitation. Default
is \code{0.001}.}

\item{p0_rainyday}{A numeric value for the probability of precipitation following
a rainy day. Default is \code{0.001}.}

\item{p0_dryday}{A numeric value for the probability of precipitation following
a dry day. Default is \code{0.001}.}

\item{set_seed}{An optional numerical value to set a seed for the analysis. Default \code{NULL}.}

\item{return_type}{A string denoting whether to return the generated rainfall as a
vector or in the data frame.}
}
\value{
A data frame containing the infilled rainfall data, including columns
for station name, date, RFE, original rainfall, and generated rainfall.
}
\description{
This function infills missing or incomplete rainfall data based on specified
parameters, statistical distributions, and historical calibration data. It
uses monthly parameters, supports a Markov chain approach, and allows for
user-defined dry seasons and custom bins for rainfall estimates.
}
\examples{

\dontrun{
# Example with the Zambia data set
data("zambia_data")

infill_data <- do_infilling(data = zambia_data,
                            station = "station",
                            date = "date",
                            rainfall = "rainfall",
                            rfe = "rfe",
                            lon = "lon",
                            lat = "lat",
                            stations_to_include = "PETAUKE MET",
                            rainfall_estimate_column = "chirps")
}

}
\seealso{
\code{\link{weather_generator}}, \code{\link{compute_monthly_parameters}}
station to exclude to become stations to include (and amend code accordingly). Default everything()
creates rfe column and selects columns to include/exclude in later calculations
Model/adjustments defined here
}
