% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rows_by_date_range_across_years.R
\name{extract_rows_by_date_range_across_years}
\alias{extract_rows_by_date_range_across_years}
\title{Extract Rows by Date Range Across Multiple Years}
\usage{
extract_rows_by_date_range_across_years(df, date, start_date, end_date)
}
\arguments{
\item{df}{A data frame containing the \code{date} variable}

\item{date}{A string specifying the column name in \code{df} that contains
the date values.}

\item{start_date}{A string in "MM-DD" format specifying the start of the date range.}

\item{end_date}{A string in "MM-DD" format specifying the end of the date range.}
}
\value{
A filtered data frame containing rows where the \code{date} falls within the specified date range, across all years.
}
\description{
This function filters rows in a data frame based on a specified date range, ignoring the year
component and treating the date range as recurring annually. Useful for selecting seasonal data.
}
