% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_quantiles.R
\name{get_n_quantiles}
\alias{get_n_quantiles}
\title{Get specified number of quantiles from a variable.}
\usage{
get_n_quantiles(x, n_quantiles, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{n_quantiles}{The number of quantiles to calculate.}

\item{na.rm}{A logical value indicating whether NA values should be removed before computation. Defaults to FALSE.}
}
\value{
A named numeric vector of quantiles.
}
\description{
This function calculates and returns the specified number of quantiles of a numeric vector.
}
\examples{
data <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
get_n_quantiles(data, 4) # Quartiles
get_n_quantiles(data, 5) # Quintiles
get_n_quantiles(c(1,2,NA,3,4), 4, na.rm = TRUE)
}
