% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_function.R
\name{logistic_function}
\alias{logistic_function}
\title{Logistic Function}
\usage{
logistic_function(rfe, b0, b1)
}
\arguments{
\item{rfe}{A numeric value or vector representing the rainfall estimate (RFE).}

\item{b0}{A numeric value representing the intercept of the logistic model.}

\item{b1}{A numeric value representing the slope of the logistic model.}
}
\value{
A numeric value or vector giving the probability output from the logistic function.
}
\description{
Computes the logistic function for a given RFE value and parameters.
This function can be used to estimate conditional probabilities using a logistic model.
}
