% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_generator.R
\name{weather_generator}
\alias{weather_generator}
\title{Generate Synthetic Rainfall Data}
\usage{
weather_generator(
  data,
  station,
  date,
  rfe,
  rainfall,
  metadata = NULL,
  metadata_station = NULL,
  lon,
  lat,
  monthly_params_df,
  distribution_flag = "gamma",
  markovflag = TRUE
)
}
\arguments{
\item{data}{Either a path to a CSV file containing historical rainfall data
or a data frame.}

\item{station}{A string specifying the column name in \code{data} that contains
the station names.}

\item{date}{A string specifying the column name in \code{data} that contains
the date values.}

\item{rfe}{A string specifying the column name in \code{data} that contains
rainfall estimates (RFE) used as predictors for rainfall generation.}

\item{rainfall}{A string specifying the column name in \code{data} that contains
the original rainfall values.}

\item{metadata}{An optional data frame containing additional metadata to merge
with the historical data. Default is \code{NULL}.}

\item{metadata_station}{A string specifying the column name in the metadata
data frame that corresponds to \code{station}.
If \code{NULL}, it defaults to the value of \code{station}.}

\item{lon}{A string specifying the column corresponding to longitude values.
If \code{metadata} is \code{NULL}, this column is from \code{data}, otherwise
this is from \code{metadata}}

\item{lat}{A string specifying the column corresponding to latitude values.
If \code{metadata} is \code{NULL}, this column is from \code{data}, otherwise
this is from \code{metadata}}

\item{monthly_params_df}{A data frame containing monthly parameters for
rainfall generation. This must include columns
for \code{Month}, \code{b0}, \code{b1}, \code{b0_rainyday}, \code{b1_rainyday},
\code{b0_dryday}, \code{b1_dryday}, \code{a0}, \code{a1}, \code{kappa},
\code{theta}, \code{p0}, \code{p0_rainyday}, and \code{p0_dryday}.}

\item{distribution_flag}{A string specifying the statistical distribution
to use for generating rainfall amounts. Options
are \code{"gamma"} (default) or \code{"lognormal"}.}

\item{markovflag}{A logical value indicating whether to use a Markov chain
approach for rainfall occurrence. If \code{TRUE} (default), the
probability of rainfall depends on previous conditions.}
}
\value{
A data frame containing the generated synthetic rainfall data with
the following columns:
- \code{station_col}: The station name.
- \code{date}: The date.
- \code{lon}: Longitude of the station.
- \code{lat}: Latitude of the station.
- \code{rfe}: Rainfall estimates (RFE) used in the generation process.
- \code{original_rainfall}: The original rainfall value from \code{data}.
- \code{generated_rainfall}: The generated synthetic rainfall value.
}
\description{
This function generates synthetic rainfall data using historical rainfall
data and monthly parameters. It applies a specified statistical distribution
(gamma or lognormal) to model rainfall amounts, optionally using a Markov
chain to simulate the probability of precipitation based on past conditions.
}
