% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zambia_data.R
\docType{data}
\name{zambia_data}
\alias{zambia_data}
\title{Zambia Meteorological Data}
\format{
A data frame with several rows and 15 variables:
\describe{
\item{station}{The name of the meteorological station. There are five stations in this data: PETAUKE MET, MSEKERA AGROMET, MFUWE MET, LUNDAZI MET and CHIPATA MET.}
\item{date}{Date of observation from 1st January 1983 until 31st December 2022. Format: YYYY-MM-DD.}
\item{lon}{Longitude of the station's location, in degrees.}
\item{lat}{Latitude of the station's location, in degrees.}
\item{rainfall}{Measured rainfall in millimeters.}
\item{rfe}{Rainfall estimates.}
\item{chirps}{Rainfall estimate from CHIRPS (Climate Hazards Group InfraRed Precipitation with Station data), in millimeters.}
\item{imerg_cal}{Calibrated rainfall estimates from IMERG (Integrated Multi-satellite Retrievals for GPM), in millimeters.}
\item{imerg_uncal}{Uncalibrated rainfall estimates from IMERG, in millimeters.}
\item{era5}{Rainfall data derived from ERA5 (ECMWF Reanalysis 5th Generation), in millimeters.}
\item{uwnd_925}{U-component of wind (zonal wind) at 925 hPa, in meters per second.}
\item{uwnd_600}{U-component of wind (zonal wind) at 600 hPa, in meters per second.}
\item{vwnd_925}{V-component of wind (meridional wind) at 925 hPa, in meters per second.}
\item{vwnd_600}{V-component of wind (meridional wind) at 600 hPa, in meters per second.}
\item{tamsat}{Rainfall estimates from TAMSAT (Tropical Applications of Meteorology using SATellite data), in millimeters.}
}
}
\source{
Data collected from five Meteorological Stations in Zambia and derived from multiple meteorological sources.
}
\usage{
zambia_data
}
\description{
This dataset contains weather and climate-related variables recorded at Meteorological Stations in Zambia. The data spans multiple dates, stations, and provides key meteorological and model-derived information.
}
\keyword{datasets}
