% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_monthly_parameters.R
\name{compute_monthly_parameters}
\alias{compute_monthly_parameters}
\title{Compute Monthly Parameters}
\usage{
compute_monthly_parameters(
  data,
  date,
  custom_bins = c(1, 3, 5, 10, 15),
  count_filter = 10,
  min_rainy_days_threshold = 10
)
}
\arguments{
\item{data}{Either a path to a CSV file containing historical rainfall data
or a data frame.}

\item{date}{A string specifying the column name in \code{data} that contains
the date values.}

\item{custom_bins}{A numeric vector specifying custom bins for RFE.}

\item{count_filter}{A numeric threshold defining the minimum number of values in each bin to include in the calculations}

\item{min_rainy_days_threshold}{A numeric threshold for minimum rainy days.}
}
\value{
A data frame with the computed monthly parameters.
}
\description{
Calculates monthly parameters based on rainfall data from a CSV file.
}
