% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_monthly_parameters.R
\name{compute_monthly_parameters}
\alias{compute_monthly_parameters}
\title{Compute Monthly Parameters}
\usage{
compute_monthly_parameters(
  filein,
  custom_bins = c(1, 3, 5, 10, 15),
  count_filter = 10,
  min_rainy_days_threshold = 10
)
}
\arguments{
\item{filein}{A string representing the input CSV file path.}

\item{custom_bins}{A numeric vector specifying custom bins for RFE.}

\item{count_filter}{A numeric threshold for filtering counts.}

\item{min_rainy_days_threshold}{A numeric threshold for minimum rainy days.}
}
\value{
A data frame with the computed monthly parameters.
}
\description{
Calculates monthly parameters based on rainfall data from a CSV file.
}
