% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_calibration_data.R
\name{select_calibration_data}
\alias{select_calibration_data}
\title{Select Calibration Data}
\usage{
select_calibration_data(
  data,
  station,
  rainfall_estimate_column,
  stations_to_include = NULL,
  save = FALSE
)
}
\arguments{
\item{data}{Either a path to a CSV file containing historical rainfall data
or a data frame.}

\item{station}{A string specifying the column name in \code{data} that contains
the station names.}

\item{rainfall_estimate_column}{A string representing the column name with rainfall estimates to be used in calibration.}

\item{stations_to_include}{A string representing the name of the station to exclude from the calibration data (Optional).}

\item{save}{A logical column indicating whether to save the resulting data frame or not. Default \code{FALSE}.}
}
\value{
A data frame with the calibration data (without rows from the excluded station).
}
\description{
This function reads a CSV file containing rainfall data, selects the relevant
column for rainfall estimates, and excludes data from a specified station for calibration.
It outputs a CSV file with the filtered calibration data.
}
