% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_generator.R
\name{weather_generator}
\alias{weather_generator}
\title{Generate Synthetic Rainfall Data}
\usage{
weather_generator(
  datain,
  monthly_params_df,
  distribution_flag = "gamma",
  markovflag = 1
)
}
\arguments{
\item{datain}{Either a path to a CSV file containing historical rainfall data
or a data frame. The data frame must include columns for
\code{Station_name}, \code{date}, \code{rfe}, \code{rainfall}, \code{lon}, and \code{lat}.}

\item{monthly_params_df}{A data frame containing monthly parameters for
generating rainfall data. It should include columns
for \code{Month}, \code{b0}, \code{b1}, \code{b0_rainyday}, \code{b1_rainyday},
\code{b0_dryday}, \code{b1_dryday}, \code{a0}, \code{a1}, \code{kappa},
\code{theta}, \code{p0}, \code{p0_rainyday}, and \code{p0_dryday}.}

\item{distribution_flag}{A string indicating the distribution to use for
generating rainfall amounts. Options are 'gamma'
(default) or 'lognormal'.}

\item{markovflag}{A binary indicator (default 1) for whether to use a Markov
chain approach in determining the probability of rainfall.
If set to 0, the function does not consider previous
rainfall when calculating probabilities.}
}
\value{
A data frame containing the generated rainfall data, including columns
for \code{Station_name}, \code{date}, \code{lon}, \code{lat}, \code{rfe}, \code{original_rainfall},
and \code{generated_rainfall}.
}
\description{
This function generates synthetic rainfall data based on historical rainfall
events and specified monthly parameters. It uses a specified statistical
distribution (gamma or lognormal) to model the rainfall amounts.
}
\examples{
# Example usage:
monthly_params <- data.frame(Month = 1:12, b0 = runif(12), b1 = runif(12), 
                              b0_rainyday = runif(12), b1_rainyday = runif(12),
                              b0_dryday = runif(12), b1_dryday = runif(12),
                              a0 = runif(12), a1 = runif(12), kappa = runif(12),
                              theta = runif(12), p0 = runif(12),
                              p0_rainyday = runif(12), p0_dryday = runif(12))
historical_data <- data.frame(Station_name = "Station A", date = Sys.Date(),
                               rfe = c(0, 1, 0), rainfall = c(0, 5, 0),
                               lon = 34.0, lat = -1.0)
generated_rainfall <- weather_generator(historical_data, monthly_params)
print(generated_rainfall)

}
